Changed layout, added % touch damage, and elements are listed better now (thanks to Spadow)

+++ WzComparerR2.MapRender/LifeInfo.cs
@@ -14,7 +14,7 @@ namespace WzComparerR2.MapRender
 
         public int level;
         public long maxHP;
-        public int maxMP;
+        public long maxMP;
         public int speed;
         public int PADamage;
         public int PDDamage;
@@ -67,7 +67,7 @@ namespace WzComparerR2.MapRender
                     {
                         case "level": lifeInfo.level = node.GetValueEx<int>(0); break;
                         case "maxHP": lifeInfo.maxHP = node.GetValueEx<long>(0); break;
-                        case "maxMP": lifeInfo.maxMP = node.GetValueEx<int>(0); break;
+                        case "maxMP": lifeInfo.maxMP = node.GetValueEx<long>(0); break;
                         case "speed": lifeInfo.speed = node.GetValueEx<int>(0); break;
                         case "PADamage": lifeInfo.PADamage = node.GetValueEx<int>(0); break;
                         case "PDDamage": lifeInfo.PDDamage = node.GetValueEx<int>(0); break;
